Abstracted styles from cytoscape.js
