+++ src/lib/components/cytoscape/Cytoscape.svelte
@@ -22,6 +22,7 @@
     import {page} from "$app/state";
     import {pushState} from "$app/navigation";
     import { fetchCourseData, getEdgeData, getNodeData } from "./Data";
+    import { getStyles } from "./Styles";
 
     let focus = $derived(page.url.searchParams.get('focus'));
 
@@ -101,10 +102,6 @@
             pushState(page.url, page.state);
         }
     })
-
-    type StyleData = {
-        [parent: string]: string;
-    };
 
     let progress = $state({
         text: "Loading Graph...",
@@ -191,90 +188,7 @@
             number: 50,
         }
 
-        let styleResponse = await fetch(styleUrl);
-        let styleData: StyleData[] = await styleResponse.json();
-
-        let cytoscapeStyles: StylesheetStyle[] = [
-            {
-                selector: 'node',
-                style: {
-                    'label': 'data(id)',
-                    'text-valign': 'center',
-                    'text-halign': 'center',
-                    'background-color': '#757575',
-                }
-            },
-            {
-                selector: '.highlighted-nodes',
-                style: {
-                    'border-width': 1,
-                    'border-color': '#000',
-                    'border-style': 'solid',
-                }
-            },
-            {
-                selector: '.faded',
-                style: {
-                    'opacity': 0.25,
-                    'text-opacity': 0.25,
-                }
-            },
-            {
-                selector: '*',
-                style: {
-                    'transition-property': 'opacity',
-                    'transition-duration': 0.2,
-                }
-            },
-            {
-                selector: 'node[type="compound"]',
-                style: {
-                    'text-valign': 'top',
-                    'border-width': 0,
-                    'background-opacity': 0,
-                    label: '',
-                }
-            },
-            {
-                selector: 'edge',
-                style: {
-                    'width': 1,
-                    'line-color': '#000',
-                    'curve-style': 'straight',
-                    'target-arrow-color': '#000',
-                    'target-arrow-shape': 'triangle',
-                    'source-distance-from-node': 5,
-                    'target-distance-from-node': 5,
-                    'text-wrap': 'wrap',
-                    'font-size': 10,
-                }
-            },
-            {
-                selector: '.highlighted-edges',
-                style: {
-                    'width': 2,
-                }
-            },
-            {
-                selector: '.no-overlay',
-                style: {
-                    'overlay-padding': 0,
-                    'overlay-opacity': 0,
-                }
-            }
-        ]
-
-        const styles = styleData.map(item => {
-            const [parent, color] = Object.entries(item)[0];
-            return {
-                selector: `node[parent="${parent}"]`,
-                style: {
-                    'background-color': color,
-                },
-            };
-        });
-
-        cytoscapeStyles = cytoscapeStyles.concat(styles);
+        let cytoscapeStyles = getStyles(styleUrl);
 
         progress = {
             text: "Loading Layout...",
